refactor : Passive 리팩토링
