+++ Assets/Resources/Prefabs/Passive.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 9004e3d69b785b4449b86d6c4f72d302
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
