+++ Assets/Resources/Prefabs/Passive/Passive.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: f761dbc7f270e5b47ba72b219100bd83
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
