+++ Assets/Resources/Prefabs/Passive/가학.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 11db3f0537fb35543a931f72a4cdeb3d
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
