+++ Assets/Resources/Prefabs/Passive/강림.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: a6a73fff7fbbe034eb24cf4eea75874e
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
