+++ Assets/Resources/Prefabs/Passive/고문관.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 8ed302a24519aec4aa627fe3bcff55fe
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
