+++ Assets/Resources/Prefabs/Passive/고양.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 67ed1f89a0f9d3e43990f1d07a09087a
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
