+++ Assets/Resources/Prefabs/Passive/대죄.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 6b7375150ca85564fa7b3fc98463d79b
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
