+++ Assets/Resources/Prefabs/Passive/동생.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 5083759a370061b48a9d767b7e66171b
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
