+++ Assets/Resources/Prefabs/Passive/망령.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 29cc327993ec5354b87b4576db922bd4
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
