+++ Assets/Resources/Prefabs/Passive/오빠.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 3d7138347d0d26949be8bc0706b57230
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
