+++ Assets/Resources/Prefabs/Passive/자애.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 5156b58e1438ec345bf1db69a947a266
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
