+++ Assets/Resources/Prefabs/Passive/처형.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: a8e9d8c1f281b1f498040c204d472c54
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
