+++ Assets/Resources/Prefabs/Passive/흡수.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 561726d5fee601e468f72fd8e65860cd
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
