+++ Assets/Script/Unit/Passvie.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: bc2f1211e991f3d4684b0e8a918c8b55
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
