+++ Assets/Script/Unit/Passvie/가학.cs
@@ -1,0 +1,19 @@
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+public class 가학 : Passive
+{
+    private bool isApplied = false;
+
+    public override void Use(BattleUnit caster, BattleUnit receiver)
+    {
+        if (isApplied)
+            return;
+
+        caster.ChangedStat.ATK += 3;
+        isApplied = true;
+    }
+}
