+++ Assets/Script/Unit/Passvie/가학.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 6acf9519051e83d49af5f113a45d47c1
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
