+++ Assets/Script/Unit/Passvie/강림.cs
@@ -1,0 +1,22 @@
+using System.Collections.Generic;
+using UnityEngine;
+
+public class 강림 : Passive
+{
+    public override void Use(BattleUnit caster, BattleUnit receiver)
+    {
+        List<Vector2> targetCoords = new List<Vector2>();
+        targetCoords.Add(caster.Location + Vector2.up);
+        targetCoords.Add(caster.Location + Vector2.down);
+        targetCoords.Add(caster.Location + Vector2.right);
+        targetCoords.Add(caster.Location + Vector2.left);
+
+        List<BattleUnit> targetUnits = BattleManager.Instance.GetArroundUnits(targetCoords);
+
+        foreach (BattleUnit unit in targetUnits)
+        {
+            if (unit.Team != caster.Team)
+                unit.ChangeHP(-15);
+        }
+    }
+}
