+++ Assets/Script/Unit/Passvie/강림.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 70379c90130f2374197bdf16048eee17
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
