+++ Assets/Script/Unit/Passvie/고문관.cs
@@ -1,0 +1,16 @@
+using System.Collections;
+using UnityEngine;
+
+public class 고문관 : Passive
+{
+    public override void Use(BattleUnit caster, BattleUnit receiver)
+    {
+        Vector2 moveVec = (receiver.Location - caster.Location).normalized;
+
+        if (!BattleManager.Field.TileDict[caster.Location + moveVec].UnitExist)
+        {
+            Debug.Log(caster.Location + moveVec);
+            BattleManager.Field.MoveUnit(receiver.Location, caster.Location + moveVec);
+        }
+    }
+}
