+++ Assets/Script/Unit/Passvie/고문관.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: e80051e26f4c5cc44871b0bf56b2776b
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
