+++ Assets/Script/Unit/Passvie/고양.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 632b2ab8d162f7c4b80cc0aabe051cf7
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
