+++ Assets/Script/Unit/Passvie/대죄.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: c8f49176e63b2ee40bc03f65ba3a36df
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
