+++ Assets/Script/Unit/Passvie/동생.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 2aa3d07844582f3428a589a7d6456aae
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
