+++ Assets/Script/Unit/Passvie/망령.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 279b75f04594a96478d612ffbee1108d
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
