+++ Assets/Script/Unit/Passvie/오빠.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 5b08ffc7351fd7b47bf49545719644e8
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
