+++ Assets/Script/Unit/Passvie/자애.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 797486a4715ea534982dd8e310e9990d
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
