+++ Assets/Script/Unit/Passvie/처형.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 251f87a3cf76c3741be81f1480b2383c
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
