+++ Assets/Script/Unit/Passvie/흡수.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 5dd74037f95a500419c97346f2d4dbbb
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
